Changelog: - Added new scoreboard system to improve performances - Localization fixes - long values handling in various commands - Added possibility to reset titles

+++ src/main/java/me/kaotich00/easyranking/command/admin/ScoreCommand.java
@@ -76,19 +76,19 @@ public class ScoreCommand extends ERAdminCommand {
         switch(scoreOperator) {
             case "add":
                 totalScore = boardService.addScoreToPlayer(board, playerUUID, score);
-                sender.sendMessage(ChatFormatter.formatSuccessMessage("Successfully added " + ChatColor.GOLD + ChatFormatter.thousandSeparator(score.intValue()) + " " + ChatColor.GREEN + board.getUserScoreName() + " to " + ChatColor.GOLD + playerName));
+                sender.sendMessage(ChatFormatter.formatSuccessMessage("Successfully added " + ChatColor.GOLD + ChatFormatter.thousandSeparator(score.longValue()) + " " + ChatColor.GREEN + board.getUserScoreName() + " to " + ChatColor.GOLD + playerName));
                 break;
             case "subtract":
                 totalScore = boardService.subtractScoreFromPlayer(board, playerUUID, score);
-                sender.sendMessage(ChatFormatter.formatSuccessMessage("Successfully subtracted " + ChatColor.GOLD + ChatFormatter.thousandSeparator(score.intValue()) + " " + ChatColor.GREEN + board.getUserScoreName() + " from " + ChatColor.GOLD + playerName));
+                sender.sendMessage(ChatFormatter.formatSuccessMessage("Successfully subtracted " + ChatColor.GOLD + ChatFormatter.thousandSeparator(score.longValue()) + " " + ChatColor.GREEN + board.getUserScoreName() + " from " + ChatColor.GOLD + playerName));
                 break;
             case "set":
                 totalScore = boardService.setScoreOfPlayer(board, playerUUID, score);
-                sender.sendMessage(ChatFormatter.formatSuccessMessage("Successfully set " + ChatColor.GOLD + ChatFormatter.thousandSeparator(score.intValue()) + " " + ChatColor.GREEN + board.getUserScoreName() + " to " + ChatColor.GOLD + playerName));
+                sender.sendMessage(ChatFormatter.formatSuccessMessage("Successfully set " + ChatColor.GOLD + ChatFormatter.thousandSeparator(score.longValue()) + " " + ChatColor.GREEN + board.getUserScoreName() + " to " + ChatColor.GOLD + playerName));
                 break;
         }
 
-        sender.sendMessage(ChatFormatter.formatSuccessMessage(ChatColor.GRAY + "New score for " + ChatColor.GOLD + playerName + ChatColor.GRAY + ": " + ChatColor.GREEN + ChatFormatter.thousandSeparator(totalScore.intValue())));
+        sender.sendMessage(ChatFormatter.formatSuccessMessage(ChatColor.GRAY + "New score for " + ChatColor.GOLD + playerName + ChatColor.GRAY + ": " + ChatColor.GREEN + ChatFormatter.thousandSeparator(totalScore.longValue())));
         return;
     }
